Implement fixed-point FMA and testbench
By default, the FMA sets its result to its private register for c so it
can perform chained dot products. There are valid_in signals to reset a,
b, and c. The "compute" flag tells the FMA when to compute.

The FMA testbenches to work at one operation per cycle!