Implement and testbench FMA (#3)
By default, the FMA sets its result to its private register for c so that it
can perform chained dot products. There are valid_in signals to reset a,
b, and c. The "compute" flag tells the FMA when to compute.

The FMA works at one operation per cycle in the testbench!