Use program file directly [bug fix]
Because SystemVerilog processes macros before it processes parameters,
passing in the program name as a parameter to the controller module does
not work. Passing the program file in directly allows the BRAM to read
it.

Now the ISA successfully runs the all-no-ops.mem program at one
instruction per two cycles. Later, it would be possible to use
prefetching to achieve one instruction per cycle. However, this
optimization might not be important given that wait times for the data
cache may ultimately dominate latency.